+++ jitb/jitb_globals.py
@@ -7,6 +7,10 @@ from typing import Final, List
 # Default {'role': 'system', 'content': DEFAULT_SYSTEM_CONTENT}
 DEFAULT_SYSTEM_CONTENT: Final[str] = 'You are a funny person trying to win Jackbox Games.'
 
+# Argument parser supported "commands"
+JITB_ARG_CMDS_AUTO: Final[List[str]] = ['automatic', 'auto']              # Auto commands
+JITB_ARG_CMDS_MAN: Final[List[str]] = ['manual', 'man']                   # Man commands
+JITB_ARG_CMDS: Final[List[str]] = JITB_ARG_CMDS_AUTO + JITB_ARG_CMDS_MAN  # All commands
 JITB_POLL_RATE: Final[float] = 0.5  # Rate, in seconds, JITB will parse page content
 
 # List of Character accessible names for the Jackbox Games Quiplash 3 avatars