+++ jitb/jitb_main.py
@@ -17,25 +17,23 @@ def main() -> int:
         0 on success, 1 for failure.
     """
     # LOCAL VARIABLES
-    exit_code = 0   # 0 for success, 1 for failure.
-    room_code = ''  # Jackbox Games room code
-    username = ''   # Jackbox Games username
-    debug = False   # Debug command line argument
-    client = None   # JitbAi object
+    exit_code = 0    # 0 for success, 1 for failure.
+    client = None    # JitbAi object
+    arg_vals = None  # ArgVals object
 
     # DO IT
-    (room_code, username, debug) = parse_args()
+    arg_vals = parse_args()
     try:
-        Logger.initialize(debugging=debug)
+        Logger.initialize(debugging=arg_vals.debug)
         client = JitbAi(temperature=1.0)
         client.setup()
-        play_the_game(room_code=room_code, username=username, ai_obj=client)
+        play_the_game(room_code=arg_vals.room_code, username=arg_vals.username, ai_obj=client)
     except Exception as err:
         _print_exception(err)
         exit_code = 1
     finally:
         Logger.shutdown()
-        if debug:
+        if arg_vals.debug:
             input('[DEBUG] Game is over.  If there is an Exception, consider saving the log and '
                   'webpage for testing.  Press [Enter] to exit.')
         if client: