+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for manual login (e.g., Twitch-enabled games)
+
 ### Changed
+
+- Argument parser now utilizes commands: automatic, manual
 
 ### Deprecated